+++ packages/connectivity/src/scp-cf/cache.spec.ts
@@ -27,27 +27,26 @@ const cacheOne = new Cache<Destination>({ hours: 0, minutes: 5, seconds: 0 });
 const cacheTwo = new Cache<ClientCredentialsResponse>();
 
 describe('Cache', () => {
-  afterEach(() => {
+  afterEach(async () => {
     cacheOne.clear();
     cacheTwo.clear();
-    destinationCache.clear();
+    await destinationCache.clear();
     clientCredentialsTokenCache.clear();
   });
 
-  it('non-existing item in cache should return undefined', () => {
-    const actual = cacheOne.get('notExistingDest');
-    expect(actual).toBeUndefined();
+  it('non-existing item in cache should return undefined', async () => {
+    expect(cacheOne.get('notExistingDest')).toBeUndefined();
   });
 
   it('item should be retrieved correctly', () => {
-    cacheOne.set('one', destinationOne);
+    cacheOne.set('one', { entry: destinationOne });
     const actual = cacheOne.get('one');
     expect(actual).toEqual(destinationOne);
   });
 
   it('retrieving expired item should return undefined', () => {
     jest.useFakeTimers('modern');
-    cacheOne.set('one', destinationOne);
+    cacheOne.set('one', { entry: destinationOne });
 
     const minutesToExpire = 6;
     // Shift time to expire the set item
@@ -56,7 +55,7 @@ describe('Cache', () => {
   });
 
   it('clear() should remove all entries in cache', () => {
-    cacheOne.set('one', destinationOne);
+    cacheOne.set('one', { entry: destinationOne });
     cacheOne.clear();
     expect(cacheOne.hasKey('one')).toBeFalsy();
   });
@@ -69,7 +68,7 @@ describe('Cache', () => {
       jti: '',
       scope: ''
     };
-    cacheTwo.set('someToken', dummyToken);
+    cacheTwo.set('someToken', { entry: dummyToken });
     expect(cacheTwo.get('someToken')).toEqual(dummyToken);
   });
 
@@ -81,14 +80,17 @@ describe('Cache', () => {
       jti: '',
       scope: ''
     };
-    cacheTwo.set('expiredToken', dummyToken, 10);
-    cacheTwo.set('validToken', dummyToken, Date.now() + 5000);
+    cacheTwo.set('expiredToken', { entry: dummyToken, expires: 10 });
+    cacheTwo.set('validToken', {
+      entry: dummyToken,
+      expires: Date.now() + 5000
+    });
     expect(cacheTwo.get('expiredToken')).toBeUndefined();
     expect(cacheTwo.get('validToken')).toBe(dummyToken);
   });
 
   it('should not hit cache for undefined key', () => {
-    cacheOne.set(undefined, {} as Destination);
+    cacheOne.set(undefined, { entry: {} as Destination });
     const actual = cacheOne.get(undefined);
     expect(actual).toBeUndefined();
   });