+++ packages/connectivity/src/scp-cf/client-credentials-token-cache.ts
@@ -18,11 +18,12 @@ const ClientCredentialsTokenCache = (
     clientId: string,
     token: ClientCredentialsResponse
   ): void => {
-    cache.set(
-      getCacheKey(url, clientId),
-      token,
-      token.expires_in ? Date.now() + token.expires_in * 1000 : undefined
-    );
+    cache.set(getCacheKey(url, clientId), {
+      entry: token,
+      expires: token.expires_in
+        ? Date.now() + token.expires_in * 1000
+        : undefined
+    });
   },
   clear: (): void => {
     cache.clear();