+++ test-packages/test-services-odata-v2/test-service/TestNestedComplexType.ts
@@ -66,7 +66,7 @@ export class TestNestedComplexTypeField<
    * Representation of the [[TestNestedComplexType.stringProperty]] property for query construction.
    * Use to reference this property in query operations such as 'filter' in the fluent request API.
    */
-  stringProperty: EdmTypeField<
+  stringProperty: OrderableEdmTypeField<
     EntityT,
     DeSerializersT,
     'Edm.String',