+++ test-packages/test-services-odata-v4/test-service/TestComplexType.d.ts
@@ -168,7 +168,7 @@ export declare class TestComplexTypeField<
    * Representation of the [[TestComplexType.stringProperty]] property for query construction.
    * Use to reference this property in query operations such as 'filter' in the fluent request API.
    */
-  stringProperty: EdmTypeField<
+  stringProperty: OrderableEdmTypeField<
     EntityT,
     DeSerializersT,
     'Edm.String',
@@ -179,7 +179,7 @@ export declare class TestComplexTypeField<
    * Representation of the [[TestComplexType.booleanProperty]] property for query construction.
    * Use to reference this property in query operations such as 'filter' in the fluent request API.
    */
-  booleanProperty: EdmTypeField<
+  booleanProperty: OrderableEdmTypeField<
     EntityT,
     DeSerializersT,
     'Edm.Boolean',
@@ -190,7 +190,13 @@ export declare class TestComplexTypeField<
    * Representation of the [[TestComplexType.guidProperty]] property for query construction.
    * Use to reference this property in query operations such as 'filter' in the fluent request API.
    */
-  guidProperty: EdmTypeField<EntityT, DeSerializersT, 'Edm.Guid', true, false>;
+  guidProperty: OrderableEdmTypeField<
+    EntityT,
+    DeSerializersT,
+    'Edm.Guid',
+    true,
+    false
+  >;
   /**
    * Representation of the [[TestComplexType.int16Property]] property for query construction.
    * Use to reference this property in query operations such as 'filter' in the fluent request API.
@@ -397,7 +403,7 @@ export declare class TestComplexTypeField<
    * Representation of the [[TestComplexType.baseStringProperty]] property for query construction.
    * Use to reference this property in query operations such as 'filter' in the fluent request API.
    */
-  baseStringProperty: EdmTypeField<
+  baseStringProperty: OrderableEdmTypeField<
     EntityT,
     DeSerializersT,
     'Edm.String',