+++ test-packages/test-services-odata-v4/test-service/TestEntityApi.d.ts
@@ -14,8 +14,8 @@ import {
   EntityBuilderType,
   EntityApi,
   FieldBuilder,
-  EdmTypeField,
   OrderableEdmTypeField,
+  EdmTypeField,
   CollectionField,
   EnumField,
   OneToManyLink,
@@ -49,7 +49,7 @@ export declare class TestEntityApi<
   get fieldBuilder(): FieldBuilder<typeof TestEntity, DeSerializersT>;
   private _schema?;
   get schema(): {
-    KEY_PROPERTY_GUID: EdmTypeField<
+    KEY_PROPERTY_GUID: OrderableEdmTypeField<
       TestEntity<
         DeSerializers<
           any,
@@ -78,7 +78,7 @@ export declare class TestEntityApi<
       false,
       true
     >;
-    KEY_PROPERTY_STRING: EdmTypeField<
+    KEY_PROPERTY_STRING: OrderableEdmTypeField<
       TestEntity<
         DeSerializers<
           any,
@@ -107,7 +107,7 @@ export declare class TestEntityApi<
       false,
       true
     >;
-    STRING_PROPERTY: EdmTypeField<
+    STRING_PROPERTY: OrderableEdmTypeField<
       TestEntity<
         DeSerializers<
           any,
@@ -136,7 +136,7 @@ export declare class TestEntityApi<
       true,
       true
     >;
-    BOOLEAN_PROPERTY: EdmTypeField<
+    BOOLEAN_PROPERTY: OrderableEdmTypeField<
       TestEntity<
         DeSerializers<
           any,
@@ -165,7 +165,7 @@ export declare class TestEntityApi<
       true,
       true
     >;
-    GUID_PROPERTY: EdmTypeField<
+    GUID_PROPERTY: OrderableEdmTypeField<
       TestEntity<
         DeSerializers<
           any,