+++ test-packages/test-services-odata-v4/test-service/TestEntitySingleLinkApi.ts
@@ -17,7 +17,6 @@ import {
   EntityBuilderType,
   EntityApi,
   FieldBuilder,
-  EdmTypeField,
   OrderableEdmTypeField,
   OneToManyLink,
   OneToOneLink
@@ -110,21 +109,21 @@ export class TestEntitySingleLinkApi<
   }
 
   private _schema?: {
-    STRING_PROPERTY: EdmTypeField<
+    STRING_PROPERTY: OrderableEdmTypeField<
       TestEntitySingleLink<DeSerializers>,
       DeSerializersT,
       'Edm.String',
       true,
       true
     >;
-    BOOLEAN_PROPERTY: EdmTypeField<
+    BOOLEAN_PROPERTY: OrderableEdmTypeField<
       TestEntitySingleLink<DeSerializers>,
       DeSerializersT,
       'Edm.Boolean',
       true,
       true
     >;
-    GUID_PROPERTY: EdmTypeField<
+    GUID_PROPERTY: OrderableEdmTypeField<
       TestEntitySingleLink<DeSerializers>,
       DeSerializersT,
       'Edm.Guid',
@@ -138,7 +137,7 @@ export class TestEntitySingleLinkApi<
       true,
       true
     >;
-    KEY_PROPERTY: EdmTypeField<
+    KEY_PROPERTY: OrderableEdmTypeField<
       TestEntitySingleLink<DeSerializers>,
       DeSerializersT,
       'Edm.String',