+++ packages/odata-common/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @sap-cloud-sdk/odata-common
 
+## 2.5.0
+
+### Minor Changes
+
+- c3166ff6: [Compatibility Note] Change `Edm.String`, `Edm.Boolean` and `Edm.Guid` to be orderable to support `lt`/`lessThan()`, `le`/`lessOrEqual()`, `gt`/`greaterThan()`, and `ge`/`greaterOrEqual` operators. Re-generate odata services to adopt the changes.
+
+### Patch Changes
+
+- Updated dependencies [89f1c423]
+- Updated dependencies [9481ec69]
+  - @sap-cloud-sdk/util@2.5.0
+  - @sap-cloud-sdk/http-client@2.5.0
+  - @sap-cloud-sdk/connectivity@2.5.0
+
 ## 2.4.0
 
 ### Patch Changes