+++ test-packages/test-services-odata-v2/CHANGELOG.md
@@ -1,3 +1,5 @@
 # @sap-cloud-sdk/test-services-odata-v2
 
+## 2.5.0
+
 ## 2.4.0