+++ test-packages/type-tests/CHANGELOG.md
@@ -1,3 +1,5 @@
 # @sap-cloud-sdk/type-tests
 
+## 2.5.0
+
 ## 2.4.0